+++ CloudHarvestCoreDataModel/__register__.py
@@ -1,1 +1,1 @@
-from tasks import RecordsetTask
+from tasks import HarvestRecordSetTask