+++ CloudHarvestCoreDataModel/tasks.py
@@ -5,9 +5,9 @@ from typing import List
 
 
 @register_definition(name='recordset')
-class RecordsetTask(BaseTask):
+class HarvestRecordSetTask(BaseTask):
     """
-    The RecordsetTask class is a subclass of the BaseTask class. It represents a task that operates on a record set.
+    The HarvestRecordSetTask class is a subclass of the BaseTask class. It represents a task that operates on a record set.
 
     Attributes:
         recordset_name (HarvestRecordSet): The name of the record set this task operates on.
@@ -27,7 +27,7 @@ class RecordsetTask(BaseTask):
 
     def __init__(self, recordset_name: HarvestRecordSet, stages: List[dict], *args, **kwargs):
         """
-        Constructs a new RecordsetTask instance.
+        Constructs a new HarvestRecordSetTask instance.
 
         Args:
             recordset_name (HarvestRecordSet): The name of the record set this task operates on.
@@ -49,10 +49,10 @@ class RecordsetTask(BaseTask):
         It then checks if the function is a method of the HarvestRecordSet or HarvestRecord class. If it is, it applies the function to the record set or each record in the record set, respectively.
         If the function is not a method of either class, it raises an AttributeError.
 
-        The result of applying the function is stored in the data attribute of the RecordsetTask instance.
+        The result of applying the function is stored in the data attribute of the HarvestRecordSetTask instance.
 
         Returns:
-            self: Returns the instance of the RecordsetTask.
+            self: Returns the instance of the HarvestRecordSetTask.
         """
 
         from .record import HarvestRecord
@@ -67,14 +67,15 @@ class RecordsetTask(BaseTask):
 
             # Each dictionary should only contain one key-value pair
             for function, arguments in stage.items():
+
                 # This is a HarvestRecordSet command
                 if hasattr(HarvestRecordSet, function):
-                    getattr(recordset, function)(**arguments)
+                    getattr(recordset, function)(**arguments or {})
 
                 # This is a HarvestRecord command
                 elif hasattr(HarvestRecord, function):
                     [
-                        getattr(record, function)(**arguments)
+                        getattr(record, function)(**arguments or {})
                         for record in recordset
                     ]